--- src/Components/Dashboard/Customers.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export class Customers extends React.Component {
-	render() {
-		return <h1>Customers</h1>
-	}
-}
