--- src/Components/Dashboard/Orders.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export class Orders extends React.Component {
-  render() {
-    return <h1>Orders</h1>
-  }
-}
